requirements.txt
torch>=1.12.0
numpy>=1.21.0
scipy>=1.7.0
addict>=2.4.0
pyyaml>=6.0
timm>=0.6.0
einops>=0.6.0

# PTv3 dependencies (install separately per platform)
# pip install spconv-cu118  (or spconv-cu121)
# pip install torch-scatter -f https://data.pyg.org/whl/torch-2.0.0+cu118.html

# Optional: FlashAttention (recommended for PTv3)
# pip install flash-attn --no-build-isolation

# AdaPoinTr dependencies
# pip install "git+https://github.com/erikwijmans/Pointnet2_PyTorch.git#egg=pointnet2_ops&subdirectory=pointnet2_ops_lib"

# Data generation (dataset_generator.py)
trimesh>=3.20.0
laspy>=2.3.0
rtree>=1.0.0

# Optional
open3d>=0.17.0
